+++ lib/types.ts
@@ -1,0 +1,8 @@
+import { StaticImageData } from "next/image"
+
+export type ProjectCardProps = {
+    title:string,
+    description:string,
+    tags:string[],
+    imageUrl:StaticImageData,
+}